refactor(currency-util): organize code for readability
